Refactor container structure
