+++ src/config/auth.ts
@@ -1,6 +1,6 @@
 export default {
   jwt: {
-    secret: process.env.APP_SECRET,
+    secret: process.env.APP_SECRET || '',
     expiresIn: '1h',
   },
 };
