+++ src/shared/container/providers/MailTemplate/index.ts
@@ -1,0 +1,6 @@
+import container from '../../Container';
+
+import HandlebarsMailTemplate from './implementations/HandlebarsMailTemplate';
+import IMailTemplate from './models/IMailTemplate';
+
+container.registry<IMailTemplate>('MailTemplate', () => new HandlebarsMailTemplate());
