+++ src/shared/container/providers/StorageProvider/index.ts
@@ -1,0 +1,6 @@
+import container from '../../Container';
+
+import IStorageProvider from './models/IStorageProvider';
+import DiskStorageProvider from './implementations/DiskStorageProvider';
+
+container.registry<IStorageProvider>('DiskStorage', () => new DiskStorageProvider());
